@@ -13,7 +13,7 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Data
 	#region usings
 
 	using System.Collections.Generic;
-	using Newtonsoft.Json;
+	using System.Text.Json.Serialization;
 
 	#endregion
 
@@ -50,7 +50,8 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Data
 		/// when the inspection plan search is performed with the versioning flag set. When creating new parts,
 		/// this information will be ignored by the server.
 		/// </summary>
-		[JsonProperty( "history" )]
+		[Newtonsoft.Json.JsonProperty( "history" )]
+		[JsonPropertyName( "history" )]
 		public IReadOnlyList<InspectionPlanCharacteristicDto> History { get; set; }
 
 		#endregion
