@@ -12,8 +12,7 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Data
 {
 	#region usings
 
-	using Newtonsoft.Json;
-	using Newtonsoft.Json.Converters;
+	using System.Text.Json.Serialization;
 
 	#endregion
 
@@ -22,7 +21,8 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Data
 	/// beim Koppeln von Messungen verschiedener Teile anhand eines primären Messungsschlüssels
 	/// für jeden im Suchergebnis enthaltenen Attributwert des Schlüssels erfüllt sein muss.
 	/// </summary>
-	[JsonConverter( typeof( StringEnumConverter ) )]
+	[Newtonsoft.Json.JsonConverter( typeof( Newtonsoft.Json.Converters.StringEnumConverter ) )]
+	[JsonConverter( typeof( JsonStringEnumConverter ) )]
 	public enum MeasurementMergeConditionDto
 	{
 		/// <summary>
