@@ -14,7 +14,7 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Data
 
 	using System;
 	using System.Collections.Generic;
-	using Newtonsoft.Json;
+	using System.Text.Json.Serialization;
 
 	#endregion
 
@@ -33,88 +33,103 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Data
 		/// <summary>
 		/// Gets or sets the server name.
 		/// </summary>
-		[JsonProperty( "serverName" )]
+		[Newtonsoft.Json.JsonProperty( "serverName" )]
+		[JsonPropertyName( "serverName" )]
 		public string ServerName { get; set; }
 
 		/// <summary>
 		/// Gets or sets the server version.
 		/// </summary>
-		[JsonProperty( "version" )]
+		[Newtonsoft.Json.JsonProperty( "version" )]
+		[JsonPropertyName( "version" )]
 		public string Version { get; set; }
 
 		/// <summary>
 		/// Gets or sets whether the server has security enabled.
 		/// </summary>
-		[JsonProperty( "securityEnabled" )]
+		[Newtonsoft.Json.JsonProperty( "securityEnabled" )]
+		[JsonPropertyName( "securityEnabled" )]
 		public bool SecurityEnabled { get; set; }
 
 		/// <summary>
 		/// Gets or sets the servers edition.
 		/// </summary>
-		[JsonProperty( "edition" )]
+		[Newtonsoft.Json.JsonProperty( "edition" )]
+		[JsonPropertyName( "edition" )]
 		public string Edition { get; set; }
 
 		/// <summary>
 		/// Gets or sets the number of parts that currently exist in the server. This number is just an approximation.
 		/// </summary>
-		[JsonProperty( "partCount" )]
+		[Newtonsoft.Json.JsonProperty( "partCount" )]
+		[JsonPropertyName( "partCount" )]
 		public int PartCount { get; set; }
 
 		/// <summary>
 		/// Gets or sets the number of characteristics that currently exist in the server. This number is just an approximation.
 		/// </summary>
-		[JsonProperty( "characteristicCount" )]
+		[Newtonsoft.Json.JsonProperty( "characteristicCount" )]
+		[JsonPropertyName( "characteristicCount" )]
 		public int CharacteristicCount { get; set; }
 
 		/// <summary>
 		/// Gets or sets the number of measurements that currently exist in the server. This number is just an approximation.
 		/// </summary>
-		[JsonProperty( "measurementCount" )]
+		[Newtonsoft.Json.JsonProperty( "measurementCount" )]
+		[JsonPropertyName( "measurementCount" )]
 		public int MeasurementCount { get; set; }
 
 		/// <summary>
 		/// Gets or sets the number of values that currently exist in the server. This number is just an approximation.
 		/// </summary>
-		[JsonProperty( "valueCount" )]
+		[Newtonsoft.Json.JsonProperty( "valueCount" )]
+		[JsonPropertyName( "valueCount" )]
 		public int ValueCount { get; set; }
 
 		/// <summary>
 		/// Gets or sets a list of features that are supported by the server.
 		/// </summary>
-		[JsonProperty( "featureList" )]
+		[Newtonsoft.Json.JsonProperty( "featureList" )]
+		[JsonPropertyName( "featureList" )]
 		public IReadOnlyCollection<string> FeatureList { get; set; }
 
 		/// <summary>
 		/// Gets or sets the timestamp of the last inspection plan modification accross the whole server.
 		/// </summary>
-		[JsonProperty( "inspectionPlanTimestamp" )]
+		[Newtonsoft.Json.JsonProperty( "inspectionPlanTimestamp" )]
+		[JsonPropertyName( "inspectionPlanTimestamp" )]
 		public DateTime? InspectionPlanTimestamp { get; set; }
 
 		/// <summary>
 		/// Gets or sets the timestamp of the last structure modification accross the whole server.
 		/// </summary>
-		[JsonProperty( "structureTimestamp" )]
+		[Newtonsoft.Json.JsonProperty( "structureTimestamp" )]
+		[JsonPropertyName( "structureTimestamp" )]
 		public DateTime? StructureTimestamp { get; set; }
 
 		/// <summary>
 		/// Gets or sets the timestamp of the last measurement modification accross the whole server.
 		/// </summary>
-		[JsonProperty( "measurementTimestamp" )]
+		[Newtonsoft.Json.JsonProperty( "measurementTimestamp" )]
+		[JsonPropertyName( "measurementTimestamp" )]
 		public DateTime? MeasurementTimestamp { get; set; }
 
 		/// <summary>
 		/// Gets or sets the timestamp for the last modification of the server configuration.
 		/// </summary>
-		[JsonProperty( "configurationTimestamp" )]
+		[Newtonsoft.Json.JsonProperty( "configurationTimestamp" )]
+		[JsonPropertyName( "configurationTimestamp" )]
 		public DateTime? ConfigurationTimestamp { get; set; }
 
 		/// <summary>
 		/// Gets or sets the timestamp for the last modification of the catalogs.
 		/// </summary>
-		[JsonProperty( "catalogTimestamp" )]
+		[Newtonsoft.Json.JsonProperty( "catalogTimestamp" )]
+		[JsonPropertyName( "catalogTimestamp" )]
 		public DateTime? CatalogTimestamp { get; set; }
 
-		[JsonProperty( "requestHeaderSize" )]
+		[Newtonsoft.Json.JsonProperty( "requestHeaderSize" )]
+		[JsonPropertyName( "requestHeaderSize" )]
 		public int RequestHeaderSize { get; set; }
 
 		#endregion
