@@ -13,7 +13,7 @@ namespace Zeiss.PiWeb.Api.Rest.HttpClient.Data
 	#region usings
 
 	using System;
-	using Newtonsoft.Json;
+	using System.Text.Json.Serialization;
 
 	#endregion
 
@@ -25,31 +25,36 @@ namespace Zeiss.PiWeb.Api.Rest.HttpClient.Data
 		/// <summary>
 		/// Gets or sets the id of this version entry.
 		/// </summary>
-		[JsonProperty( "id" )]
+		[Newtonsoft.Json.JsonProperty( "id" )]
+		[JsonPropertyName( "id" )]
 		public int ID { get; set; }
 
 		/// <summary>
 		/// Gets or sets the versioning comment.
 		/// </summary>
-		[JsonProperty( "comment" )]
+		[Newtonsoft.Json.JsonProperty( "comment" )]
+		[JsonPropertyName( "comment" )]
 		public string Comment { get; set; }
 
 		/// <summary>
 		/// Gets or sets the comment of this inspection plan entity.
 		/// </summary>
-		[JsonProperty( "user" )]
+		[Newtonsoft.Json.JsonProperty( "user" )]
+		[JsonPropertyName( "user" )]
 		public string User { get; set; }
 
 		/// <summary>
 		/// Gets or sets the number of changed inspection plan items that where change.
 		/// </summary>
-		[JsonProperty( "changeCount" )]
+		[Newtonsoft.Json.JsonProperty( "changeCount" )]
+		[JsonPropertyName( "changeCount" )]
 		public int ChangeCount { get; set; }
 
 		/// <summary>
 		/// Contains the date and time of the versioning change.
 		/// </summary>
-		[JsonProperty( "dateCreated" )]
+		[Newtonsoft.Json.JsonProperty( "dateCreated" )]
+		[JsonPropertyName( "dateCreated" )]
 		public DateTime DateCreated { get; set; }
 
 		#endregion
