@@ -36,7 +36,7 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Converter
 		public override object ReadJson( JsonReader reader, Type objectType, object existingValue, JsonSerializer serializer )
 		{
 			var key = default( ushort );
-			var length = default( ushort );
+			var length = default( ushort? );
 			var queryEfficient = default( bool );
 			var description = default( string );
 			var attributeDefinitionType = default( string );
