@@ -16,11 +16,12 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Data
 	using System.Collections.Generic;
 	using System.Data.SqlTypes;
 	using System.Diagnostics;
+	using System.Text.Json.Serialization;
 	using System.Xml;
 	using JetBrains.Annotations;
-	using Newtonsoft.Json;
 	using Zeiss.PiWeb.Api.Definitions;
 	using Zeiss.PiWeb.Api.Rest.Dtos.Converter;
+	using Zeiss.PiWeb.Api.Rest.Dtos.JsonConverters;
 
 	#endregion
 
@@ -46,13 +47,15 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Data
 		/// <summary>
 		/// Gets or sets the uuid of this measurement.
 		/// </summary>
-		[JsonProperty( "uuid" )]
+		[Newtonsoft.Json.JsonProperty( "uuid" )]
+		[JsonPropertyName( "uuid" )]
 		public Guid Uuid { get; set; }
 
 		/// <summary>
 		/// Gets or sets the uuid the part this measurement belongs to.
 		/// </summary>
-		[JsonProperty( "partUuid" )]
+		[Newtonsoft.Json.JsonProperty( "partUuid" )]
+		[JsonPropertyName( "partUuid" )]
 		public Guid PartUuid { get; set; }
 
 		/// <summary>
@@ -60,26 +63,30 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Data
 		/// timestamp whenever an attribute of this measurement is changed or whenever measurement values
 		/// of this measurement are updated, deleted and added.
 		/// </summary>
-		[JsonProperty( "lastModified" )]
+		[Newtonsoft.Json.JsonProperty( "lastModified" )]
+		[JsonPropertyName( "lastModified" )]
 		public DateTime LastModified { get; set; }
 
 		/// <summary>
 		/// Gets or sets the creation timestamp of this measurement.
 		/// </summary>
-		[JsonProperty( "created" )]
+		[Newtonsoft.Json.JsonProperty( "created" )]
+		[JsonPropertyName( "created" )]
 		public DateTime Created { get; set; }
 
 		/// <summary>
 		/// Gets or sets the status information for this measurement. This status information can be requested when
 		/// performing a measurement search using one of the values from <see cref="MeasurementStatisticsDto"/>.
 		/// </summary>
-		[JsonProperty( "status" )]
+		[Newtonsoft.Json.JsonProperty( "status" )]
+		[JsonPropertyName( "status" )]
 		public IReadOnlyList<SimpleMeasurementStatusDto> Status { get; set; }
 
 		/// <summary>
 		/// Gets or sets the time of this measurement. If this measurement has no time attribute, then <see cref="MinimumValidDatabaseDateTime"/> will
 		/// be returned.
 		/// </summary>
+		[Newtonsoft.Json.JsonIgnore]
 		[JsonIgnore]
 		public DateTime? Time
 		{
@@ -124,12 +131,14 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Data
 		/// <summary>
 		/// Returns the measurement time and in case of no time is specified, the minimum time allowed (<see cref="System.Data.SqlTypes.SqlDateTime.MinValue"/>).
 		/// </summary>
+		[Newtonsoft.Json.JsonIgnore]
 		[JsonIgnore]
 		public DateTime TimeOrMinDate => Time ?? MinimumValidDatabaseDateTime;
 
 		/// <summary>
 		/// Returns the measurement time and in case of no time is specified, the creation date of the measurement.
 		/// </summary>
+		[Newtonsoft.Json.JsonIgnore]
 		[JsonIgnore]
 		public DateTime TimeOrCreationDate => Time ?? Created;
 
@@ -148,7 +157,8 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Data
 		#region interface IAttributeItemDto
 
 		/// <inheritdoc />
-		[JsonProperty( "attributes" ), JsonConverter( typeof( AttributeArrayConverter ) )]
+		[Newtonsoft.Json.JsonProperty( "attributes" ), Newtonsoft.Json.JsonConverter( typeof( AttributeArrayConverter ) )]
+		[JsonPropertyName( "attributes" ), JsonConverter( typeof( JsonAttributeArrayConverter ) )]
 		public IReadOnlyList<AttributeDto> Attributes
 		{
 			[NotNull] get => _Attributes;
