@@ -13,8 +13,8 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.RawData
 	#region usings
 
 	using System;
+	using System.Text.Json.Serialization;
 	using JetBrains.Annotations;
-	using Newtonsoft.Json;
 
 	#endregion
 
@@ -72,14 +72,16 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.RawData
 		/// <summary>
 		/// Gets or sets the target object this raw data object belongs to.
 		/// </summary>
-		[JsonProperty( "target" )]
+		[Newtonsoft.Json.JsonProperty( "target" )]
+		[JsonPropertyName( "target" )]
 		public RawDataTargetEntityDto Target { get; set; }
 
 		/// <summary>
 		/// This is a unique key that identifies this specific raw data object for a corresponding entity. An entity (Part, Characteristic, Measurement, Value)
 		/// can have multiple raw data object that are distinct by this key.
 		/// </summary>
-		[JsonProperty( "key" )]
+		[Newtonsoft.Json.JsonProperty( "key" )]
+		[JsonPropertyName( "key" )]
 		public int? Key { get; set; }
 
 		/// <summary>
@@ -88,13 +90,15 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.RawData
 		/// <remarks>
 		/// Please note that this filename is not unique (unlike filenames in traditional file systems).
 		/// </remarks>
-		[JsonProperty( "fileName" )]
+		[Newtonsoft.Json.JsonProperty( "fileName" )]
+		[JsonPropertyName( "fileName" )]
 		public string FileName { get; set; }
 
 		/// <summary>
 		/// Gets or sets the mime type of the raw data object.
 		/// </summary>
-		[JsonProperty( "mimeType" )]
+		[Newtonsoft.Json.JsonProperty( "mimeType" )]
+		[JsonPropertyName( "mimeType" )]
 		public string MimeType { get; set; }
 
 		/// <summary>
@@ -104,7 +108,8 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.RawData
 		/// The attribute <code>LastModified</code> will be set by the Rawdata-Service.
 		/// A user value will be ignored.
 		/// </remarks>
-		[JsonProperty( "lastModified" )]
+		[Newtonsoft.Json.JsonProperty( "lastModified" )]
+		[JsonPropertyName( "lastModified" )]
 		public DateTime LastModified { get; set; }
 
 		/// <summary>
@@ -114,19 +119,22 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.RawData
 		/// The attribute <code>Created</code> will be set by the Rawdata-Service.
 		/// A user value will be ignored.
 		/// </remarks>
-		[JsonProperty( "created" )]
+		[Newtonsoft.Json.JsonProperty( "created" )]
+		[JsonPropertyName( "created" )]
 		public DateTime Created { get; set; }
 
 		/// <summary>
 		/// Gets or sets the size of the raw data object in bytes.
 		/// </summary>
-		[JsonProperty( "size" )]
+		[Newtonsoft.Json.JsonProperty( "size" )]
+		[JsonPropertyName( "size" )]
 		public int Size { get; set; }
 
 		/// <summary>
 		/// Gets or sets the MD5-Hash of the raw data object.
 		/// </summary>
-		[JsonProperty( "md5" )]
+		[Newtonsoft.Json.JsonProperty( "md5" )]
+		[JsonPropertyName( "md5" )]
 		public Guid MD5 { get; set; }
 
 		/// <summary>
@@ -135,6 +143,7 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.RawData
 		/// <remarks>
 		/// This property exists for serialization compatibility reasons only. This will be removed in a future version of this interface.
 		/// </remarks>
+		[Newtonsoft.Json.JsonIgnore]
 		[JsonIgnore]
 		public string MD5String
 		{
