@@ -20,8 +20,8 @@ namespace Zeiss.PiWeb.Api.Rest.Common.Client
 	using System.Net.Http;
 	using System.Net.Http.Headers;
 	using System.Text;
+	using System.Threading;
 	using JetBrains.Annotations;
-	using Zeiss.PiWeb.Api.Rest.Common.Data;
 	using Zeiss.PiWeb.Api.Rest.Dtos;
 
 	#endregion
@@ -35,7 +35,7 @@ namespace Zeiss.PiWeb.Api.Rest.Common.Client
 		/// </summary>
 		public static Func<HttpRequestMessage> CreateGet( string relativeUri )
 		{
-			return Create( HttpMethod.Get, relativeUri, Payload.Empty, null, Array.Empty<ParameterDefinition>() );
+			return Create( HttpMethod.Get, relativeUri, null, Array.Empty<ParameterDefinition>() );
 		}
 
 		/// <summary>
@@ -44,7 +44,7 @@ namespace Zeiss.PiWeb.Api.Rest.Common.Client
 		/// </summary>
 		public static Func<HttpRequestMessage> CreateGet( string relativeUri, ParameterDefinition parameterDefinition )
 		{
-			return Create( HttpMethod.Get, relativeUri, Payload.Empty, null, new[] { parameterDefinition } );
+			return Create( HttpMethod.Get, relativeUri, null, new[] { parameterDefinition } );
 		}
 
 		/// <summary>
@@ -53,7 +53,7 @@ namespace Zeiss.PiWeb.Api.Rest.Common.Client
 		/// </summary>
 		public static Func<HttpRequestMessage> CreateGet( string relativeUri, IEnumerable<ParameterDefinition> parameterDefinitions )
 		{
-			return Create( HttpMethod.Get, relativeUri, Payload.Empty, null, parameterDefinitions );
+			return Create( HttpMethod.Get, relativeUri, null, parameterDefinitions );
 		}
 
 		/// <summary>
@@ -61,7 +61,7 @@ namespace Zeiss.PiWeb.Api.Rest.Common.Client
 		/// </summary>
 		public static Func<HttpRequestMessage> CreateDelete( string relativeUri )
 		{
-			return Create( HttpMethod.Delete, relativeUri, Payload.Empty, null, Array.Empty<ParameterDefinition>() );
+			return Create( HttpMethod.Delete, relativeUri, null, Array.Empty<ParameterDefinition>() );
 		}
 
 		/// <summary>
@@ -70,7 +70,7 @@ namespace Zeiss.PiWeb.Api.Rest.Common.Client
 		/// </summary>
 		public static Func<HttpRequestMessage> CreateDelete( string relativeUri, ParameterDefinition parameterDefinition )
 		{
-			return Create( HttpMethod.Delete, relativeUri, Payload.Empty, null, new[] { parameterDefinition } );
+			return Create( HttpMethod.Delete, relativeUri, null, new[] { parameterDefinition } );
 		}
 
 		/// <summary>
@@ -79,14 +79,14 @@ namespace Zeiss.PiWeb.Api.Rest.Common.Client
 		/// </summary>
 		public static Func<HttpRequestMessage> CreateDelete( string relativeUri, IEnumerable<ParameterDefinition> parameterDefinitions )
 		{
-			return Create( HttpMethod.Delete, relativeUri, Payload.Empty, null, parameterDefinitions );
+			return Create( HttpMethod.Delete, relativeUri, null, parameterDefinitions );
 		}
 
 		/// <summary>
 		/// Creates a new POST-<see cref="HttpRequestMessage"/> based on the <paramref name="relativeUri"/>.
 		/// The body of the HTTP message is provided by the <paramref name="payload"/> parameter.
 		/// </summary>
-		public static Func<HttpRequestMessage> CreatePost( string relativeUri, [NotNull] Payload payload )
+		public static Func<IObjectSerializer, CancellationToken, HttpRequestMessage> CreatePost( string relativeUri, [NotNull] Payload payload )
 		{
 			return Create( HttpMethod.Post, relativeUri, payload, null, Array.Empty<ParameterDefinition>() );
 		}
@@ -96,7 +96,7 @@ namespace Zeiss.PiWeb.Api.Rest.Common.Client
 		/// and extended by possible additional query parameters represented by <paramref name="parameterDefinition"/>
 		/// The body of the HTTP message is provided by the <paramref name="payload"/> parameter.
 		/// </summary>
-		public static Func<HttpRequestMessage> CreatePost( string relativeUri, [NotNull] Payload payload, ParameterDefinition parameterDefinition )
+		public static Func<IObjectSerializer, CancellationToken, HttpRequestMessage> CreatePost( string relativeUri, [NotNull] Payload payload, ParameterDefinition parameterDefinition )
 		{
 			return Create( HttpMethod.Post, relativeUri, payload, null, new [] { parameterDefinition } );
 		}
@@ -106,7 +106,7 @@ namespace Zeiss.PiWeb.Api.Rest.Common.Client
 		/// and extended by possible additional query parameters represented by <paramref name="parameterDefinitions"/>
 		/// The body of the HTTP message is provided by the <paramref name="payload"/> parameter.
 		/// </summary>
-		public static Func<HttpRequestMessage> CreatePost( string relativeUri, [NotNull] Payload payload, IEnumerable<ParameterDefinition> parameterDefinitions )
+		public static Func<IObjectSerializer, CancellationToken, HttpRequestMessage> CreatePost( string relativeUri, [NotNull] Payload payload, IEnumerable<ParameterDefinition> parameterDefinitions )
 		{
 			return Create( HttpMethod.Post, relativeUri, payload, null, parameterDefinitions );
 		}
@@ -115,7 +115,7 @@ namespace Zeiss.PiWeb.Api.Rest.Common.Client
 		/// Creates a new PUT-<see cref="HttpRequestMessage"/> based on the <paramref name="relativeUri"/>.
 		/// The body of the HTTP message is provided by the <paramref name="payload"/> parameter.
 		/// </summary>
-		public static Func<HttpRequestMessage> CreatePut( string relativeUri, [NotNull] Payload payload )
+		public static Func<IObjectSerializer, CancellationToken, HttpRequestMessage> CreatePut( string relativeUri, [NotNull] Payload payload )
 		{
 			return Create( HttpMethod.Put, relativeUri, payload, null, Array.Empty<ParameterDefinition>() );
 		}
@@ -125,7 +125,7 @@ namespace Zeiss.PiWeb.Api.Rest.Common.Client
 		/// and extended by possible additional query parameters represented by <paramref name="parameterDefinition"/>
 		/// The body of the HTTP message is provided by the <paramref name="payload"/> parameter.
 		/// </summary>
-		public static Func<HttpRequestMessage> CreatePut( string relativeUri, [NotNull] Payload payload, ParameterDefinition parameterDefinition )
+		public static Func<IObjectSerializer, CancellationToken, HttpRequestMessage> CreatePut( string relativeUri, [NotNull] Payload payload, ParameterDefinition parameterDefinition )
 		{
 			return Create( HttpMethod.Put, relativeUri, payload, null, new[] { parameterDefinition } );
 		}
@@ -135,7 +135,7 @@ namespace Zeiss.PiWeb.Api.Rest.Common.Client
 		/// and extended by possible additional query parameters represented by <paramref name="parameterDefinitions"/>
 		/// The body of the HTTP message is provided by the <paramref name="payload"/> parameter.
 		/// </summary>
-		public static Func<HttpRequestMessage> CreatePut( string relativeUri, [NotNull] Payload payload, IEnumerable<ParameterDefinition> parameterDefinitions )
+		public static Func<IObjectSerializer, CancellationToken, HttpRequestMessage> CreatePut( string relativeUri, [NotNull] Payload payload, IEnumerable<ParameterDefinition> parameterDefinitions )
 		{
 			return Create( HttpMethod.Put, relativeUri, payload, null, parameterDefinitions );
 		}
@@ -146,11 +146,20 @@ namespace Zeiss.PiWeb.Api.Rest.Common.Client
 		/// The body of the HTTP message is provided by the <paramref name="payload"/> parameter.
 		/// </summary>
 		/// <exception cref="ArgumentNullException"><paramref name="payload"/> is <see langword="null" />.</exception>
-		public static Func<HttpRequestMessage> Create( HttpMethod method, string relativeUri, [NotNull] Payload payload, KeyValuePair<string, string>[] additionalHttpRequestHeader, IEnumerable<ParameterDefinition> parameterDefinitions )
+		public static Func<IObjectSerializer, CancellationToken, HttpRequestMessage> Create( HttpMethod method, string relativeUri, [NotNull] Payload payload, KeyValuePair<string, string>[] additionalHttpRequestHeader, IEnumerable<ParameterDefinition> parameterDefinitions )
 		{
 			if( payload == null ) throw new ArgumentNullException( nameof( payload ) );
 
-			return () => CreateInternal( method, relativeUri, payload, additionalHttpRequestHeader, parameterDefinitions );
+			return (serializer, cancellationToken ) => CreateInternal( method, relativeUri, payload, serializer, cancellationToken, additionalHttpRequestHeader, parameterDefinitions );
+		}
+
+		/// <summary>
+		/// Creates a new <see cref="HttpRequestMessage"/> for http verb <paramref name="method"/> based on the <paramref name="relativeUri"/>
+		/// and extended by possible additional query parameters represented by <paramref name="parameterDefinitions"/>
+		/// </summary>
+		public static Func<HttpRequestMessage> Create( HttpMethod method, string relativeUri, KeyValuePair<string, string>[] additionalHttpRequestHeader, IEnumerable<ParameterDefinition> parameterDefinitions )
+		{
+			return () => CreateInternal( method, relativeUri, Payload.Empty, null, CancellationToken.None, additionalHttpRequestHeader, parameterDefinitions );
 		}
 
 		/// <summary>
@@ -158,16 +167,14 @@ namespace Zeiss.PiWeb.Api.Rest.Common.Client
 		/// and extended by possible additional query parameters represented by <paramref name="parameterDefinitions"/>
 		/// The body of the HTTP message is provided by the <paramref name="payload"/> parameter.
 		/// </summary>
-		private static HttpRequestMessage CreateInternal( HttpMethod method, string relativeUri, Payload payload, KeyValuePair<string, string>[] additionalHttpRequestHeader, IEnumerable<ParameterDefinition> parameterDefinitions )
+		private static HttpRequestMessage CreateInternal( HttpMethod method, string relativeUri, Payload payload, IObjectSerializer serializer, CancellationToken cancellationToken, KeyValuePair<string, string>[] additionalHttpRequestHeader, IEnumerable<ParameterDefinition> parameterDefinitions )
 		{
 			var request = SetParametersAndHeaders( method, relativeUri, additionalHttpRequestHeader, parameterDefinitions );
 			if( payload != Payload.Empty )
 			{
 				request.Content = new PushStreamContent( ( outputStream, _, _ ) =>
 				{
-					using var sw = new StreamWriter( outputStream, Encoding.UTF8, 64 * 1024, false );
-					RestClientHelper.CreateJsonSerializer().Serialize( sw, payload.Value );
-
+					return serializer.SerializeAsync( outputStream, payload.Value, cancellationToken );
 				}, new MediaTypeWithQualityHeaderValue( RestClientBase.MimeTypeJson ) );
 			}
 
@@ -197,7 +204,7 @@ namespace Zeiss.PiWeb.Api.Rest.Common.Client
 		/// </summary>
 		private static HttpRequestMessage CreateWithAttachmentInternal( [NotNull] HttpMethod method, string relativeUri, Stream stream, string mimeType, long? contentLength, Guid? contentMD5, string contentDisposition, params ParameterDefinition[] parameterDefinitions )
 		{
-			var request = CreateInternal( method, relativeUri, Payload.Empty, null, parameterDefinitions );
+			var request = CreateInternal( method, relativeUri, Payload.Empty, null, CancellationToken.None, null, parameterDefinitions );
 
 			if( stream == null )
 				return request;
