@@ -29,8 +29,8 @@ namespace Zeiss.PiWeb.Api.Rest.Common.Client
 		/// Initializes a new instance of the <see cref="RestClient" /> class.
 		/// </summary>
 		/// <exception cref="ArgumentNullException"><paramref name="serverUri"/> is <see langword="null" />.</exception>
-		public RestClient( [NotNull] Uri serverUri, string endpointName, TimeSpan? timeout = null, int maxUriLength = DefaultMaxUriLength, bool chunked = true )
-			: base( serverUri, endpointName, timeout, maxUriLength, chunked )
+		public RestClient( [NotNull] Uri serverUri, string endpointName, TimeSpan? timeout = null, int maxUriLength = DefaultMaxUriLength, bool chunked = true, [CanBeNull] IObjectSerializer serializer = null )
+			: base( serverUri, endpointName, timeout, maxUriLength, chunked, serializer: serializer )
 		{ }
 
 		#endregion
