@@ -23,6 +23,7 @@ namespace Zeiss.PiWeb.Api.Rest.Common.Client
 	using System.Net;
 	using System.Net.Http;
 	using System.Net.Http.Headers;
+	using System.Runtime.CompilerServices;
 	using System.Runtime.Versioning;
 	using System.Text;
 	using System.Threading;
@@ -30,7 +31,6 @@ namespace Zeiss.PiWeb.Api.Rest.Common.Client
 	using CacheCow.Client;
 	using CacheCow.Common;
 	using JetBrains.Annotations;
-	using Newtonsoft.Json;
 	using Zeiss.PiWeb.Api.Rest.Common.Data;
 	using Zeiss.PiWeb.Api.Rest.Contracts;
 	using Zeiss.PiWeb.Api.Rest.Dtos;
@@ -83,6 +83,8 @@ namespace Zeiss.PiWeb.Api.Rest.Common.Client
 		private readonly bool _Chunked;
 		[CanBeNull] private readonly DelegatingHandler _CustomHttpMessageHandler;
 
+		private readonly IObjectSerializer _Serializer;
+
 		private HttpClient _HttpClient;
 		private TimeoutHandler _TimeoutHandler;
 		private HttpClientHandler _HttpClientHandler;
@@ -106,7 +108,8 @@ namespace Zeiss.PiWeb.Api.Rest.Common.Client
 			TimeSpan? timeout = null,
 			int maxUriLength = DefaultMaxUriLength,
 			bool chunked = true,
-			[CanBeNull] DelegatingHandler customHttpMessageHandler = null )
+			[CanBeNull] DelegatingHandler customHttpMessageHandler = null,
+			[CanBeNull] IObjectSerializer serializer = null )
 		{
 			if( serverUri == null )
 				throw new ArgumentNullException( nameof( serverUri ) );
@@ -118,6 +121,7 @@ namespace Zeiss.PiWeb.Api.Rest.Common.Client
 
 			_Chunked = chunked;
 			_CustomHttpMessageHandler = customHttpMessageHandler;
+			_Serializer = serializer ?? ObjectSerializer.Default;
 
 			MaxUriLength = maxUriLength;
 
@@ -253,9 +257,19 @@ namespace Zeiss.PiWeb.Api.Rest.Common.Client
 			return PerformRequestAsync<object>( requestCreationHandler, false, null, true, cancellationToken );
 		}
 
+		public Task Request( [NotNull] Func<IObjectSerializer, CancellationToken, HttpRequestMessage> requestCreationHandler, CancellationToken cancellationToken )
+		{
+			return PerformRequestAsync<object>( () => requestCreationHandler( _Serializer, cancellationToken ), false, null, true, cancellationToken );
+		}
+
 		public Task<T> Request<T>( [NotNull] Func<HttpRequestMessage> requestCreationHandler, CancellationToken cancellationToken )
 		{
-			return PerformRequestAsync( requestCreationHandler, false, ResponseToObjectAsync<T>, true, cancellationToken );
+			return PerformRequestAsync( requestCreationHandler, false, response => ResponseToObjectAsync<T>( response, _Serializer, cancellationToken ), true, cancellationToken );
+		}
+
+		public Task<T> Request<T>( [NotNull] Func<IObjectSerializer, CancellationToken, HttpRequestMessage> requestCreationHandler, CancellationToken cancellationToken )
+		{
+			return PerformRequestAsync( () => requestCreationHandler( _Serializer, cancellationToken ), false, response => ResponseToObjectAsync<T>( response, _Serializer, cancellationToken ), true, cancellationToken );
 		}
 
 		public Task<Stream> RequestStream( [NotNull] Func<HttpRequestMessage> requestCreationHandler, CancellationToken cancellationToken )
@@ -263,9 +277,29 @@ namespace Zeiss.PiWeb.Api.Rest.Common.Client
 			return PerformRequestAsync( requestCreationHandler, true, ResponseToStreamAsync, false, cancellationToken );
 		}
 
-		public Task<IEnumerable<T>> RequestEnumerated<T>( [NotNull] Func<HttpRequestMessage> requestCreationHandler, CancellationToken cancellationToken )
+		public Task<Stream> RequestStream( [NotNull] Func<IObjectSerializer, CancellationToken, HttpRequestMessage> requestCreationHandler, CancellationToken cancellationToken )
 		{
-			return PerformRequestAsync( requestCreationHandler, true, ResponseToEnumerationAsync<T>, false, cancellationToken );
+			return PerformRequestAsync( () => requestCreationHandler( _Serializer, cancellationToken ), true, ResponseToStreamAsync, false, cancellationToken );
+		}
+
+		public async IAsyncEnumerable<T> RequestEnumerated<T>( [NotNull] Func<HttpRequestMessage> requestCreationHandler, [EnumeratorCancellation] CancellationToken cancellationToken )
+		{
+			var items = await PerformRequestAsync( requestCreationHandler, true, response => Task.FromResult( ResponseToAsyncEnumerable<T>( response, _Serializer, cancellationToken ) ), false, cancellationToken );
+
+			await foreach( var item in items.ConfigureAwait( false ) )
+			{
+				yield return item;
+			}
+		}
+
+		public async IAsyncEnumerable<T> RequestEnumerated<T>( [NotNull] Func<IObjectSerializer, HttpRequestMessage> requestCreationHandler, [EnumeratorCancellation] CancellationToken cancellationToken )
+		{
+			var items = await PerformRequestAsync( () => requestCreationHandler( _Serializer ), true, response => Task.FromResult( ResponseToAsyncEnumerable<T>( response, _Serializer, cancellationToken ) ), false, cancellationToken );
+
+			await foreach (var item in items.ConfigureAwait( false ) )
+			{
+				yield return item;
+			}
 		}
 
 		public Task<byte[]> RequestBytes( [NotNull] Func<HttpRequestMessage> requestCreationHandler, CancellationToken cancellationToken )
@@ -273,9 +307,19 @@ namespace Zeiss.PiWeb.Api.Rest.Common.Client
 			return PerformRequestAsync( requestCreationHandler, false, ResponseToBytesAsync, true, cancellationToken );
 		}
 
+		public Task<byte[]> RequestBytes( [NotNull] Func<IObjectSerializer, HttpRequestMessage> requestCreationHandler, CancellationToken cancellationToken )
+		{
+			return PerformRequestAsync( () => requestCreationHandler( _Serializer ), false, ResponseToBytesAsync, true, cancellationToken );
+		}
+
 		public Task<T> RequestBinary<T>( [NotNull] Func<HttpRequestMessage> requestCreationHandler, CancellationToken cancellationToken )
 		{
 			return PerformRequestAsync( requestCreationHandler, false, BinaryResponseToObjectAsync<T>, true, cancellationToken );
+		}
+
+		public Task<T> RequestBinary<T>( [NotNull] Func<IObjectSerializer, HttpRequestMessage> requestCreationHandler, CancellationToken cancellationToken )
+		{
+			return PerformRequestAsync( () => requestCreationHandler( _Serializer ), false, BinaryResponseToObjectAsync<T>, true, cancellationToken );
 		}
 
 		/// <summary>
@@ -289,12 +333,28 @@ namespace Zeiss.PiWeb.Api.Rest.Common.Client
 			return PerformRequestAsync( requestCreationHandler, false, LocationHeaderToUrl, true, cancellationToken );
 		}
 
-		private static async Task<T> ResponseToObjectAsync<T>( HttpResponseMessage response )
+		/// <summary>
+		/// Start an async operation. Returns an URL to poll for status updates if the operation is accepted, otherwise the operation is done synchronously.
+		/// </summary>
+		/// <returns>Returns a Task that represents the duration of the initial REST request. The result of the task contains
+		/// the URI for polling the operation result or null in case the server already finished the request synchronously.</returns>
+		/// <exception cref="RestClientException">The response indicated status Accepted, but did not contain polling information.</exception>
+		public Task<Uri> RequestAsyncOperation( [NotNull] Func<IObjectSerializer, HttpRequestMessage> requestCreationHandler, CancellationToken cancellationToken )
 		{
-			using( var responseStream = await response.Content.ReadAsStreamAsync().ConfigureAwait( false ) )
+			return PerformRequestAsync( () => requestCreationHandler( _Serializer ), false, LocationHeaderToUrl, true, cancellationToken );
+		}
+
+		private static async Task<T> ResponseToObjectAsync<T>( HttpResponseMessage response, IObjectSerializer serializer, CancellationToken cancellationToken )
+		{
+			if( response.StatusCode != HttpStatusCode.NoContent )
 			{
-				return RestClientHelper.DeserializeObject<T>( responseStream );
+				using( var responseStream = await response.Content.ReadAsStreamAsync().ConfigureAwait( false ) )
+				{
+					return await serializer.DeserializeAsync<T>( responseStream, cancellationToken ).ConfigureAwait( false );
+				}
 			}
+
+			return default;
 		}
 
 		private static Task<Uri> LocationHeaderToUrl( HttpResponseMessage response )
@@ -306,19 +366,13 @@ namespace Zeiss.PiWeb.Api.Rest.Common.Client
 			return Task.FromResult( result );
 		}
 
-		private static async Task<IEnumerable<T>> ResponseToEnumerationAsync<T>( HttpResponseMessage response )
-		{
-			var stream = await ResponseToStreamAsync( response ).ConfigureAwait( false );
-			return GetEnumeratedResponse<T>( response, stream );
-		}
-
-		private static IEnumerable<T> GetEnumeratedResponse<T>( IDisposable response, Stream responseStream )
+		private static async IAsyncEnumerable<T> ResponseToAsyncEnumerable<T>( HttpResponseMessage response, IObjectSerializer serializer, [EnumeratorCancellation] CancellationToken cancellationToken )
 		{
 			using( response )
 			{
-				using( responseStream )
+				using( var responseStream = await ResponseToStreamAsync( response ).ConfigureAwait( false ) )
 				{
-					foreach( var item in RestClientHelper.DeserializeEnumeratedObject<T>( responseStream ) )
+					await foreach( var item in serializer.DeserializeAsyncEnumerable<T>( responseStream, cancellationToken ).ConfigureAwait( false ) )
 					{
 						yield return item;
 					}
@@ -392,7 +446,7 @@ namespace Zeiss.PiWeb.Api.Rest.Common.Client
 						continue;
 					}
 
-					await HandleFaultedResponse( response ).ConfigureAwait( false );
+					await HandleFaultedResponse( response, _Serializer, cancellationToken ).ConfigureAwait( false );
 				}
 			}
 			catch( HttpRequestException ex )
@@ -456,16 +510,16 @@ namespace Zeiss.PiWeb.Api.Rest.Common.Client
 			return response.Content.ReadAsStreamAsync();
 		}
 
-		private static async Task HandleFaultedResponse( HttpResponseMessage response )
+		private static async Task HandleFaultedResponse( HttpResponseMessage response, IObjectSerializer serializer, CancellationToken cancellationToken )
 		{
-			await HandleClientBasedFaults( response ).ConfigureAwait( false );
+			await HandleClientBasedFaults( response, serializer, cancellationToken ).ConfigureAwait( false );
 			HandleServerBasedFaults( response );
 		}
 
 		/// <summary>
 		/// Handles all responses with status codes between 400 and 499
 		/// </summary>
-		private static async Task HandleClientBasedFaults( HttpResponseMessage response )
+		private static async Task HandleClientBasedFaults( HttpResponseMessage response, IObjectSerializer serializer, CancellationToken cancellationToken )
 		{
 			if( response.StatusCode < HttpStatusCode.BadRequest || response.StatusCode >= HttpStatusCode.InternalServerError )
 				return;
@@ -475,10 +529,10 @@ namespace Zeiss.PiWeb.Api.Rest.Common.Client
 				Error error;
 				try
 				{
-					error = RestClientHelper.DeserializeObject<Error>( responseStream )
+					error = await serializer.DeserializeAsync<Error>( responseStream, cancellationToken ).ConfigureAwait( false )
 							?? new Error( $"Request {response.RequestMessage.Method} {response.RequestMessage.RequestUri} was not successful: {(int)response.StatusCode} ({response.ReasonPhrase})" );
 				}
-				catch( JsonReaderException )
+				catch( ObjectSerializerException )
 				{
 					var buffer = ( responseStream as MemoryStream )?.ToArray();
 					var content = buffer != null ? Encoding.UTF8.GetString( buffer ) : null;
@@ -556,7 +610,7 @@ namespace Zeiss.PiWeb.Api.Rest.Common.Client
 
 			// Almost all options are not available when running in browser
 			// and need to be done either manually or are not possible at all.
-			if ( !_IsBrowser )
+			if( !_IsBrowser )
 			{
 #pragma warning disable CA1416
 				_HttpClientHandler.PreAuthenticate = true;
